pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cbgb"
version = "0.1.0"
description = "CBGB (Controlled BTC Growth Bot) - Account Builder Trading System"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv==1.0.0",
    "pydantic==2.5.2",
    "structlog==23.2.0",
    "requests==2.31.0",
    "websocket-client==1.6.4",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "mypy==1.7.1",
    "ruff==0.1.6",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "plotly>=5.18.0",
    "streamlit>=1.30.0",
    "watchdog>=3.0.0",
    "pyyaml>=6.0.0",
]
dashboard = [
    "streamlit==1.30.0",
    "plotly==5.18.0",
    "watchdog==3.0.0",
    "pandas>=2.0.0",
]
analysis = [
    "pandas>=2.0.0",
    "scipy>=1.10.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "testnet: marks tests as requiring Testnet connection (deselected by default)",
]
addopts = "-m 'not testnet'"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.ruff]
line-length = 100
target-version = "py312"
